requirements.txt
asgiref==3.8.1
certifi==2024.8.30
crispy-bootstrap5==2024.2
Django==5.1
django-allauth==64.1.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
gunicorn==23.0.0
packaging==24.1
psycopg2-binary==2.9.9
python-decouple==3.8
sentry-sdk==2.13.0
setuptools==72.1.0
sqlparse==0.5.1
urllib3==2.2.2
wheel==0.43.0
whitenoise==6.7.0
